fix: edit file structure
